Extract isEnabledForWorkspace
